+++ Localization/frFR.ExtraWrathStrings.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-09-25 13-21-57 ]] --
+-- [[ Exported at 2022-10-15 17-22-05 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --